master

+++ protected/controllers/AllotmentController.php
@@ -27,7 +27,7 @@ class AllotmentController extends Controller {
         return array(
             array('allow', // allow all users to perform 'index' and 'view' actions
                 'actions' => array('index', 'view'),
-                'users' => array('*'),
+                'users' => array('@'),
 //                'roles' => array('Authenticated'),
             ),
             array('allow', // allow authenticated user to perform 'create' and 'update' actions
@@ -59,20 +59,24 @@ class AllotmentController extends Controller {
      * If creation is successful, the browser will be redirected to the 'view' page.
      */
     public function actionCreate() {
-        $model = new Allotment;
+        if (Yii::app()->user->checkAccess($this->id . '.' . $this->action->id)) {
+            $model = new Allotment;
 
 // Uncomment the following line if AJAX validation is needed
 // $this->performAjaxValidation($model);
 
-        if (isset($_POST['Allotment'])) {
-            $model->attributes = $_POST['Allotment'];
-            if ($model->save())
-                $this->redirect(array('view', 'id' => $model->id));
-        }
+            if (isset($_POST['Allotment'])) {
+                $model->attributes = $_POST['Allotment'];
+                if ($model->save())
+                    $this->redirect(array('view', 'id' => $model->id));
+            }
 
-        $this->render('create', array(
-            'model' => $model,
-        ));
+            $this->render('create', array(
+                'model' => $model,
+            ));
+        } else {
+            $this->accessDenied();
+        }
     }
 
     /**
@@ -82,19 +86,23 @@ class AllotmentController extends Controller {
      */
     public function actionUpdate($id) {
         $model = $this->loadModel($id);
+        if (Yii::app()->user->checkAccess($this->id . '.' . $this->action->id, array('owner' => $model->create_user))) {
 
 // Uncomment the following line if AJAX validation is needed
 // $this->performAjaxValidation($model);
 
-        if (isset($_POST['Allotment'])) {
-            $model->attributes = $_POST['Allotment'];
-            if ($model->save())
-                $this->redirect(array('view', 'id' => $model->id));
-        }
+            if (isset($_POST['Allotment'])) {
+                $model->attributes = $_POST['Allotment'];
+                if ($model->save())
+                    $this->redirect(array('view', 'id' => $model->id));
+            }
 
-        $this->render('update', array(
-            'model' => $model,
-        ));
+            $this->render('update', array(
+                'model' => $model,
+            ));
+        } else {
+            $this->accessDenied();
+        }
     }
 
     /**
@@ -103,16 +111,20 @@ class AllotmentController extends Controller {
      * @param integer $id the ID of the model to be deleted
      */
     public function actionDelete($id) {
-        if (Yii::app()->request->isPostRequest) {
+        if (Yii::app()->user->checkAccess($this->id . '.' . $this->action->id, array('owner' => $this->loadModel($id)->create_user))) {
+            if (Yii::app()->request->isPostRequest) {
 // we only allow deletion via POST request
-            $this->loadModel($id)->delete();
+                $this->loadModel($id)->delete();
 
 // if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
-            if (!isset($_GET['ajax']))
-                $this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+                if (!isset($_GET['ajax']))
+                    $this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+            }
+            else
+                throw new CHttpException(400, 'Invalid request. Please do not repeat this request again.');
+        }else {
+            $this->accessDenied();
         }
-        else
-            throw new CHttpException(400, 'Invalid request. Please do not repeat this request again.');
     }
 
     /**
@@ -129,14 +141,18 @@ class AllotmentController extends Controller {
      * Manages all models.
      */
     public function actionAdmin() {
-        $model = new Allotment('search');
-        $model->unsetAttributes();  // clear any default values
-        if (isset($_GET['Allotment']))
-            $model->attributes = $_GET['Allotment'];
+        if (Yii::app()->user->checkAccess($this->id . '.' . $this->action->id)) {
+            $model = new Allotment('search');
+            $model->unsetAttributes();  // clear any default values
+            if (isset($_GET['Allotment']))
+                $model->attributes = $_GET['Allotment'];
 
-        $this->render('admin', array(
-            'model' => $model,
-        ));
+            $this->render('admin', array(
+                'model' => $model,
+            ));
+        }else {
+            $this->accessDenied();
+        }
     }
 
     /**
